Refactoring commands.go into new commands package, with separate file for each subcommand
