+++ commands/commands_test.go
@@ -1,18 +1,13 @@
-package main
+package commands
 
 import (
-	"bytes"
-	"flag"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
-	"path/filepath"
-	"strings"
 	"testing"
 
-	"github.com/codegangsta/cli"
 	"github.com/docker/machine/drivers/fakedriver"
+	_ "github.com/docker/machine/drivers/none"
 	"github.com/docker/machine/libmachine"
 	"github.com/docker/machine/libmachine/auth"
 	"github.com/docker/machine/libmachine/engine"
@@ -307,291 +302,6 @@ func TestRunActionForeachMachine(t *testing.T) {
 		state, _ := machine.Driver.GetState()
 		if expected[machine.Name] != state {
 			t.Fatalf("Expected machine %s to have state %s, got state %s", machine.Name, state, expected[machine.Name])
-		}
-	}
-}
-
-func TestCmdConfig(t *testing.T) {
-	defer cleanup()
-
-	stdout := os.Stdout
-	r, w, _ := os.Pipe()
-
-	os.Stdout = w
-
-	defer func() {
-		os.Stdout = stdout
-	}()
-
-	store, err := getTestStore()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	mcn, err := libmachine.New(store)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	flags := getTestDriverFlags()
-	hostOptions := &libmachine.HostOptions{
-		EngineOptions: &engine.EngineOptions{},
-		SwarmOptions: &swarm.SwarmOptions{
-			Master:    false,
-			Discovery: "",
-			Address:   "",
-			Host:      "",
-		},
-		AuthOptions: &auth.AuthOptions{},
-	}
-
-	host, err := mcn.Create("test-a", "none", hostOptions, flags)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if err := store.SetActive(host); err != nil {
-		t.Fatalf("error setting active host: %v", err)
-	}
-
-	outStr := make(chan string)
-
-	go func() {
-		var testOutput bytes.Buffer
-		io.Copy(&testOutput, r)
-		outStr <- testOutput.String()
-	}()
-
-	set := flag.NewFlagSet("config", 0)
-	globalSet := flag.NewFlagSet("test", 0)
-	globalSet.String("storage-path", store.GetPath(), "")
-
-	c := cli.NewContext(nil, set, globalSet)
-
-	cmdConfig(c)
-
-	w.Close()
-
-	out := <-outStr
-
-	if !strings.Contains(out, "--tlsverify") {
-		t.Fatalf("Expect --tlsverify")
-	}
-
-	testMachineDir := filepath.Join(store.GetPath(), "machines", host.Name)
-
-	tlscacert := fmt.Sprintf("--tlscacert=\"%s/ca.pem\"", testMachineDir)
-	if !strings.Contains(out, tlscacert) {
-		t.Fatalf("Expected to find %s in %s", tlscacert, out)
-	}
-
-	tlscert := fmt.Sprintf("--tlscert=\"%s/cert.pem\"", testMachineDir)
-	if !strings.Contains(out, tlscert) {
-		t.Fatalf("Expected to find %s in %s", tlscert, out)
-	}
-
-	tlskey := fmt.Sprintf("--tlskey=\"%s/key.pem\"", testMachineDir)
-	if !strings.Contains(out, tlskey) {
-		t.Fatalf("Expected to find %s in %s", tlskey, out)
-	}
-
-	if !strings.Contains(out, "-H=unix:///var/run/docker.sock") {
-		t.Fatalf("Expect docker host URL")
-	}
-}
-
-func TestCmdEnvBash(t *testing.T) {
-	stdout := os.Stdout
-	shell := os.Getenv("SHELL")
-	r, w, _ := os.Pipe()
-
-	os.Stdout = w
-	os.Setenv("MACHINE_STORAGE_PATH", TestStoreDir)
-	os.Setenv("SHELL", "/bin/bash")
-
-	defer func() {
-		os.Setenv("MACHINE_STORAGE_PATH", "")
-		os.Setenv("SHELL", shell)
-		os.Stdout = stdout
-	}()
-
-	if err := clearHosts(); err != nil {
-		t.Fatal(err)
-	}
-
-	flags := getTestDriverFlags()
-
-	store, sErr := getTestStore()
-	if sErr != nil {
-		t.Fatal(sErr)
-	}
-
-	mcn, err := libmachine.New(store)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	hostOptions := &libmachine.HostOptions{
-		EngineOptions: &engine.EngineOptions{},
-		SwarmOptions: &swarm.SwarmOptions{
-			Master:    false,
-			Discovery: "",
-			Address:   "",
-			Host:      "",
-		},
-		AuthOptions: &auth.AuthOptions{},
-	}
-
-	host, err := mcn.Create("test-a", "none", hostOptions, flags)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	host, err = mcn.Get("test-a")
-	if err != nil {
-		t.Fatalf("error loading host: %v", err)
-	}
-
-	if err := mcn.SetActive(host); err != nil {
-		t.Fatalf("error setting active host: %v", err)
-	}
-
-	outStr := make(chan string)
-
-	go func() {
-		var testOutput bytes.Buffer
-		io.Copy(&testOutput, r)
-		outStr <- testOutput.String()
-	}()
-
-	set := flag.NewFlagSet("config", 0)
-	c := cli.NewContext(nil, set, set)
-	cmdEnv(c)
-
-	w.Close()
-
-	out := <-outStr
-
-	// parse the output into a map of envvar:value for easier testing below
-	envvars := make(map[string]string)
-	for _, e := range strings.Split(strings.TrimSpace(out), "\n") {
-		if !strings.HasPrefix(e, "export ") {
-			continue
-		}
-		kv := strings.SplitN(e, "=", 2)
-		key, value := kv[0], kv[1]
-		envvars[strings.Replace(key, "export ", "", 1)] = value
-	}
-
-	testMachineDir := filepath.Join(store.GetPath(), "machines", host.Name)
-
-	expected := map[string]string{
-		"DOCKER_TLS_VERIFY": "1",
-		"DOCKER_CERT_PATH":  fmt.Sprintf("\"%s\"", testMachineDir),
-		"DOCKER_HOST":       "unix:///var/run/docker.sock",
-	}
-
-	for k, v := range envvars {
-		if v != expected[k] {
-			t.Fatalf("Expected %s == <%s>, but was <%s>", k, expected[k], v)
-		}
-	}
-}
-
-func TestCmdEnvFish(t *testing.T) {
-	stdout := os.Stdout
-	shell := os.Getenv("SHELL")
-	r, w, _ := os.Pipe()
-
-	os.Stdout = w
-	os.Setenv("MACHINE_STORAGE_PATH", TestStoreDir)
-	os.Setenv("SHELL", "/bin/fish")
-
-	defer func() {
-		os.Setenv("MACHINE_STORAGE_PATH", "")
-		os.Setenv("SHELL", shell)
-		os.Stdout = stdout
-	}()
-
-	if err := clearHosts(); err != nil {
-		t.Fatal(err)
-	}
-
-	flags := getTestDriverFlags()
-
-	store, err := getTestStore()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	mcn, err := libmachine.New(store)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	hostOptions := &libmachine.HostOptions{
-		EngineOptions: &engine.EngineOptions{},
-		SwarmOptions: &swarm.SwarmOptions{
-			Master:    false,
-			Discovery: "",
-			Address:   "",
-			Host:      "",
-		},
-		AuthOptions: &auth.AuthOptions{},
-	}
-
-	host, err := mcn.Create("test-a", "none", hostOptions, flags)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	host, err = mcn.Get("test-a")
-	if err != nil {
-		t.Fatalf("error loading host: %v", err)
-	}
-
-	if err := mcn.SetActive(host); err != nil {
-		t.Fatalf("error setting active host: %v", err)
-	}
-
-	outStr := make(chan string)
-
-	go func() {
-		var testOutput bytes.Buffer
-		io.Copy(&testOutput, r)
-		outStr <- testOutput.String()
-	}()
-
-	set := flag.NewFlagSet("config", 0)
-	c := cli.NewContext(nil, set, set)
-	cmdEnv(c)
-
-	w.Close()
-
-	out := <-outStr
-
-	// parse the output into a map of envvar:value for easier testing below
-	envvars := make(map[string]string)
-	for _, e := range strings.Split(strings.TrimSuffix(out, ";\n"), ";\n") {
-		if !strings.HasPrefix(e, "set -x ") {
-			continue
-		}
-		kv := strings.SplitN(strings.Replace(e, "set -x ", "", 1), " ", 2)
-		key, value := kv[0], kv[1]
-		envvars[key] = value
-	}
-
-	testMachineDir := filepath.Join(store.GetPath(), "machines", host.Name)
-
-	expected := map[string]string{
-		"DOCKER_TLS_VERIFY": "1",
-		"DOCKER_CERT_PATH":  fmt.Sprintf("\"%s\"", testMachineDir),
-		"DOCKER_HOST":       "unix:///var/run/docker.sock",
-	}
-
-	for k, v := range envvars {
-		if v != expected[k] {
-			t.Fatalf("Expected %s == <%s>, but was <%s>", k, expected[k], v)
 		}
 	}
 }
