+++ commands/upgrade.go
@@ -1,0 +1,13 @@
+package commands
+
+import (
+	log "github.com/Sirupsen/logrus"
+
+	"github.com/codegangsta/cli"
+)
+
+func cmdUpgrade(c *cli.Context) {
+	if err := runActionWithContext("upgrade", c); err != nil {
+		log.Fatal(err)
+	}
+}
